chore: organize hooks
